Genes! Name Improvements! Code Organization! <3
This finally adds a Gene system for agents. Genes are expressed as Traits, and multiple copies of Genes can exist, be created, or deleted, with rules about their values transitioning from the old system, so functionally, the game is the same. But like, cooler. Saving/Loading fully supported, except old saves are not compatible and are no longer allowed